initial project

+++ common-module/src/main/java/tipitapi/drawmytodayimprovement/dto/PageResponse.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.function.Consumer;
 import java.util.function.Function;
 
-public record PageResponse<T>(int page, int size, int total, List<T> content) {
+public record PageResponse<T>(int page, int size, long total, List<T> content) {
 
     public <U> PageResponse<U> map(Function<? super T, U> converter) {
         return new PageResponse<>(page, size, total, getConvertedContent(converter));
